--- SonarJavaDemo/.github/workflows/sonar.yml
@@ -1,44 +1,0 @@
-name: SonarCloud
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  build:
-    name: Build and Analyze
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up JDK
-        uses: actions/setup-java@v3
-        with:
-          java-version: '11'
-          distribution: 'adopt'
-
-      - name: Cache SonarCloud packages
-        uses: actions/cache@v3
-        with:
-          path: ~/.sonar/cache
-          key: ${{ runner.os }}-sonar
-          restore-keys: ${{ runner.os }}-sonar
-
-      - name: Cache Maven packages
-        uses: actions/cache@v3
-        with:
-          path: ~/.m2
-          key: ${{ runner.os }}-maven
-          restore-keys: ${{ runner.os }}-maven
-
-      - name: Build with Maven
-        run: mvn clean verify
-
-      - name: SonarCloud Scan
-        uses: SonarSource/sonarcloud-github-action@v2
-        with:
-          organization: <ahmadk18361>
-          projectKey: <sonar-java-demo>
-          token: ${{ secrets.SONAR_TOKEN }}
