+++ sonar-project.properties
@@ -2,8 +2,8 @@ sonar.projectKey=SonarJavaDemo
 sonar.projectName=SonarJavaDemo
 sonar.projectVersion=latest
 sonar.language=java
-sonar.sources=src
-sonar.java.binaries=target
+sonar.sources=src/main/java/com/example
+sonar.java.binaries=target/classes
 sonar.java.source=17
 sonar.host.url=http://localhost:9000
 sonar.login=sqp_6e9d8c96e9fe596457705a3d718a63660286bd8a
